+++ src/main/java/team/startup/expo/domain/survey/management/repository/DynamicSurveyRepository.java
@@ -1,8 +1,8 @@
-package team.startup.expo.domain.survey.repository;
+package team.startup.expo.domain.survey.management.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import team.startup.expo.domain.survey.entity.DynamicSurvey;
-import team.startup.expo.domain.survey.entity.Survey;
+import team.startup.expo.domain.survey.management.entity.DynamicSurvey;
+import team.startup.expo.domain.survey.management.entity.Survey;
 
 import java.util.List;
