+++ src/main/java/team/startup/expo/domain/survey/management/repository/SurveyRepository.java
@@ -1,9 +1,9 @@
-package team.startup.expo.domain.survey.repository;
+package team.startup.expo.domain.survey.management.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import team.startup.expo.domain.expo.entity.Expo;
 import team.startup.expo.domain.form.entity.ParticipationType;
-import team.startup.expo.domain.survey.entity.Survey;
+import team.startup.expo.domain.survey.management.entity.Survey;
 
 import java.util.List;
 import java.util.Optional;
