+++ src/main/java/team/startup/expo/domain/survey/management/service/CreateSurveyService.java
@@ -1,0 +1,7 @@
+package team.startup.expo.domain.survey.management.service;
+
+import team.startup.expo.domain.survey.management.presentation.dto.request.SurveyRequestDto;
+
+public interface CreateSurveyService {
+    void execute(String expoId, SurveyRequestDto dto);
+}
