+++ src/main/java/team/startup/expo/domain/survey/management/service/DeleteSurveyService.java
@@ -1,4 +1,4 @@
-package team.startup.expo.domain.survey.service;
+package team.startup.expo.domain.survey.management.service;
 
 public interface DeleteSurveyService {
     void execute(String expoId);
