+++ src/main/java/team/startup/expo/domain/survey/management/service/GetSurveyService.java
@@ -1,7 +1,7 @@
-package team.startup.expo.domain.survey.service;
+package team.startup.expo.domain.survey.management.service;
 
 import team.startup.expo.domain.form.entity.ParticipationType;
-import team.startup.expo.domain.survey.presentation.dto.response.SurveyResponseDto;
+import team.startup.expo.domain.survey.management.presentation.dto.response.SurveyResponseDto;
 
 public interface GetSurveyService {
     SurveyResponseDto execute(String expoId, ParticipationType participationType);
