--- src/main/java/team/startup/expo/domain/survey/service/UpdateSurveyService.java
@@ -1,7 +1,0 @@
-package team.startup.expo.domain.survey.service;
-
-import team.startup.expo.domain.survey.presentation.dto.request.SurveyRequestDto;
-
-public interface UpdateSurveyService {
-    void execute(String expoId, SurveyRequestDto dto);
-}
